Campground Permissions

+++ routes/campgrounds.js
@@ -63,10 +63,15 @@ router.get(
     "/:id/edit",
     isLoggedIn,
     catchAsync(async (req, res) => {
-      const campground = await Campground.findById(req.params.id);
+      const { id } = req.params;
+      const campground = await Campground.findById(id);
       if (!campground) {
         req.flash('error', 'Cannot find that campground!');
         return res.redirect("/campgrounds");
+      }
+      if(!campground.author.equals(req.user._id)){
+        req.flash('error', 'You do not have permission to do that!');
+        return res.redirect(`/campgrounds/${id}`);
       }
       res.render("./campgrounds/edit", { campground });
     })
@@ -78,7 +83,12 @@ router.get(
     validateCampground,
     catchAsync(async (req, res) => {
       const { id } = req.params;
-      const campground = await Campground.findByIdAndUpdate(id, {
+      const campground = await Campground.findById(id);
+      if(!campground.author.equals(req.user._id)){
+        req.flash('error', 'You do not have permission to do that!');
+        return res.redirect(`/campgrounds/${id}`);
+      }
+      const camp = await Campground.findByIdAndUpdate(id, {
         ...req.body.campground,
       });
       req.flash('success', 'Successfully updated campground!');
@@ -91,6 +101,11 @@ router.get(
     isLoggedIn,
     catchAsync(async (req, res) => {
       const { id } = req.params;
+      const campground = await Campground.findById(id);
+      if(!campground.author.equals(req.user._id)){
+        req.flash('error', 'You do not have permission to do that!');
+        return res.redirect(`/campgrounds/${id}`);
+      }
       await Campground.findByIdAndDelete(id);
       req.flash('success', 'Successfully deleted campground!');
       res.redirect("/campgrounds");
